test: replace make-synchronous with synckit which is much faster
