+++ package.json
@@ -103,7 +103,6 @@
     "flow-bin": "0.89.0",
     "jasmine-check": "0.1.5",
     "jest": "26.6.3",
-    "make-synchronous": "^0.1.1",
     "marked": "2.1.2",
     "microtime": "3.0.0",
     "next": "11.0.1",
@@ -118,6 +117,7 @@
     "rollup-plugin-commonjs": "9.1.3",
     "rollup-plugin-json": "3.0.0",
     "rollup-plugin-strip-banner": "2.0.0",
+    "synckit": "^0.3.0",
     "transducers-js": "0.4.174",
     "tslint": "6.1.3",
     "tslint-config-prettier": "1.18.0",
